Add timeout option to cocaine-tool
Added the ability to specify a timeout on synchronous components of the
operations, when it is necessary. For example, when downloading an
application large volume or low-bandwidth network. The failure of the
action due to lack of time, you will get an appropriate error message,
therefore it is necessary to repeat the operation with a large timeout.