Storage API v2.
We decided to change the Storage API to support indexes. From now on,
storage_t::list() method is no longer available. Instead, we've provided
a new storage_t::find() method, which accepts an additional vector of
tags, which represents indexes. These tags could be added to a key with
a new optional parameter to storage_t::write().