A more reasonable default value of grow threshold.
Default value queue_limit / pool_limit * concurrency may be more than queue_limit. It leads to the fact that available workers will not be started when they are needed. I have changed the value to queue_limit / pool_limit / 2. This value makes the runtime to run all available workers when the queue is half full.