Kill JsonCpp with fire.
Whoa! It's really a big commit.
In Cocaine Core, JsonCpp was only used to read a config and Json::Value was used to store json-like data. But this library is big, it's slow, and it must be installed with the core. So i have removed it. Now cocaine::dynamic_t is used instead of Json::Value. RapidJson, which is fast header-only library, is used to parse a config.