Added configuration state machine.
'configuration_machine_t' is responsible for managing of configurations of all state machines in the cluster. When a user wants to add or remove node to some state machine, he or she should send request to this configuration machine, which will apply the change to the target state machine.
Using of the configuration machine allows to create new machines "from scratch" without system administrator help. When new node adds itself to some state machine and it is the first node in the machine (so the node creates it), the configuration machine responds, that the node is the first, and it's safe to become a leader and create new log.