Don't step down when receive vote request from unknown node.
Now Raft leader doesn't start new term, when receives vote request from node, which is not in the cluster and has old log. Such node anyway may not become a leader, and probably it's a stale node, which was removed from the cluster, but doesn't know about it. So it shouldn't force meaningless reelections.