Fixed bug in the locator's refresh method.
The refresh method doesn't update entries in the inverted index which correspond to the deleted from the routing group applications.
So when one of the applications disappears from the cloud, locator zeroes its weight in the group by an integer index, and actually it zeroes weight of some other application which actually corresponds to this integer index.