[Bug Fix] Eat connection error in an info handler.
When the engine is overloaded it can take a long time to handle an info request. The user can decide to disconnect, which results for us in "session not connected" error, after trying to write into the deferred. This exception is propagated up to event loop't thread and terminates it.

In this case here we can't do anything with it except eating an exception and to log it.