[Repo] Fixed phantom caching.
Invocation of `std::map::operator[]` creates a value-initialized object
in the map if an object with the given key was not found. In our case it
is a null weak pointer.

After that we create a shared pointer instance and insert it using
`std::map::insert`, but this method doesn't actually inserts an item if
it already exists.

The given behavior forces us to always visit the `if` branch, which is
not what we expect.

This commit should fix that bug.