[Bug Fix] Fixed deadlock when spawning.
When unable to spawn more slaves, a spawning state invokes the slave termination, which in its turn invokes self erasing from the pool. This requires acquiring a lock, so these operations mustn't be processed in a sequence.

One of the solutions is just to delay slave shutdown operation. The second one is to use slave GC instead of cleanup callbacks. The third one is to delay cleanup operation in the overseer.

This time I chose the first solution, but it leads to an infinite spinning until slave spawn.