Fix(continuum): update RGs without O(N^2) complexity.
Boost accumulate is shit.
It works through passing the accumulator value by value
each iteration. Each iteration we have linear increasing
number of copying continuums, which itself contains
hundreds of strings.

This commit forces the continuum to be move only
and applies emplace semantics for std::map with routing
groups.

For example we have 4 RG with 256, 256, 256 and 512 points
respectively. That's the result of 3 x {"app": 1} and
1 x {"app": 2, "other": 1} groups.
For this case after construction and one full update we
have ~11776 continuum `element_t` copies which involves
string reallocations.
With 8 RGs on a single refresh we suffer from ~77312
allocations and so on.

In our production there are at least 60 RGs ...

Unfortunately we can't completely get rid of copying
because of atomic RG refresh, but it's O(N).

After this commit we have linear number number of
copies on refresh.